Added XPMultiplier
This commit lets admins assign xp modifier value per user. This value is carried in the PlayerData class and can be set using setstats command. I have also added the placeholders, sql and config settings, and required messages to the plugin. When an experience is being given to the user it will be multiplied by the integer of xp multiplier. Default value will be 1